requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.4.1
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.4.0
apache-beam==2.59.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
avro-python3==1.10.2
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
catboost==1.2.7
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
courlan==1.3.2
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
dateparser==1.2.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.1.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
durationpy==0.9
easyocr==1.7.2
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.10.0
fastapi==0.115.8
fastavro==1.9.7
fasteners==0.19
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
flask-cors==5.0.1
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-auth==2.38.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
hdfs==2.7.3
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.28.1
humanfriendly==10.0
ibm-cloud-sdk-core==3.18.0
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watson_machine_learning==1.0.360
ibm_watsonx_ai==1.1.11
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.33
jaxlib==0.4.33
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
Js2Py==0.74
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jusText==3.0.1
keras==3.5.0
kiwisolver==1.4.7
kubernetes==32.0.0
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.37
langchain-huggingface==0.1.2
langchain-ibm==0.2.1
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.131
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lomond==0.3.3
lvis==0.5.3
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.1
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.8.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
object_detection==0.1
objsize==0.7.0
olefile==0.47
onnxruntime==1.20.1
openai==1.51.0
openai-whisper==20240930
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
opt-einsum==3.3.0
optree==0.14.0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.6
php==1.2.1
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
playwright==1.50.0
plotly==5.24.1
portalocker==2.10.1
posthog==3.11.0
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.29.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.5.2
pydantic_core==2.27.2
pydeck==0.9.1
pydot==1.4.2
pyee==12.1.1
Pygments==2.18.0
pyjsparser==2.7.1
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==2.4.7
pypdf==5.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.1.28
python-json-logger==2.0.7
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.12.1
redis==5.0.8
referencing==0.35.1
regex==2024.9.11
requests==2.32.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
sacrebleu==2.2.1
safetensors==0.5.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.28.1
Send2Trash==1.8.3
sentence-transformers==3.4.1
shapely==2.0.6
shellingham==1.5.4
sip==6.8.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soupsieve==2.6
split-folders==0.5.1
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.45.3
streamlit==1.39.0
streamlit-option-menu==0.4.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.18.1
tf-slim==1.1.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
tinycss2==1.3.0
tld==0.13
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.4.1
tqdm==4.66.5
trafilatura==2.0.0
traitlets==5.14.3
transformers==4.48.3
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
types-python-dateutil==2.9.0.20240906
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
unstructured==0.16.20
unstructured-client==0.29.0
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.0
watchdog==4.0.2
watchfiles==1.0.4
watson-developer-cloud==2.10.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.0.4
whois==1.20240129.2
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
wxai-langchain==0.0.4
xgboost==2.1.1
yarl==1.13.1
yfinance==0.2.43
zipp==3.20.2
zstandard==0.23.0
